refactor: move getQueryParams out of common
Avoid giving the impression that this is a canonical function or polyfill.
